requirements.txt
Pillow
tensorflow>=2.2
